+++ backport/v16/core/lib/functions.lib.php
@@ -1,8 +1,0 @@
-<?php
-
-if(!function_exists('isModEnabled')) {
-	function isModEnabled($module) {
-		global $conf;
-		return !empty($conf->$module->enabled);
-	}
-}