+++ ChangeLog.md
@@ -3,7 +3,8 @@ All notable changes to this project will be documented in this file.
 
 # [Unreleased]
 
-
+- FIX : Compat v20 
+  Changed Dolibarr compatibility range to 16 min - 20 max
 
 # Release 3.24 - 08/04/2024
 - FIX : Module_number missed in subTotal class. function  addSubTotalLine function  in test order_supplier - *26/06/2024* - 3.24.4