+++ ChangeLog.md
@@ -3,8 +3,12 @@ All notable changes to this project will be documented in this file.
 
 # [Unreleased]
 
+
+
+# Release 3.25 - 24/07/2024
 - FIX : Compat v20 
   Changed Dolibarr compatibility range to 16 min - 20 max - *11/07/2024* - 3.25.0
+- Ajout du hook pdfgeneration dans la liste des hooks accepté par le module - *12/07/2024* - 3.25.0
 
 # Release 3.24 - 08/04/2024
 - FIX : DA024845 : Le module sous total amène des erreurs dans les sauts de page lorsque l'on arrive tout juste en bas de page. - *24/07/2024* - 3.24.5